[FIX] upgrade scripts with no errors

+++ odoo/addons/base/migrations/1.4/end-migration.py
@@ -4,19 +4,12 @@
 import logging
 import os
 
-from datetime import date
-
 from openupgradelib import openupgrade
 
 _logger = logging.getLogger(__name__)
 
 
 to_remove = [
-    'mrp_eco_alphanumeric_version',
-    'mrp_kit_partial_delivery',
-    'purchase_line_invoice_line_zero',
-    'purchase_supplierinfo_currency',
-    'stock_push_scrap',
 ]
 
 
@@ -25,26 +18,18 @@ records_to_remove = [
 ]
 
 
-def change_password(env):
-    env.cr.execute("""
-        UPDATE res_users
-        SET password='admin';
-    """)
-
 @openupgrade.migrate()
 def migrate(env, installed_version):
-    _logger.warning('Delete records from XML ID')
-    openupgrade.delete_records_safely_by_xml_id(env, records_to_remove)
-    _logger.warning('Fix tax type.')
-    _logger.warning('Define company fields.')
-    _logger.warning('Update tax cash basis tax account to reconciled')
-    _logger.warning('Uninstalling not required modules')
-    modules_to_remove = env['ir.module.module'].search([
-        ('name', 'in', to_remove)])
-    modules_to_remove += modules_to_remove.downstream_dependencies()
-    modules_to_remove.module_uninstall()
-    modules_to_remove.unlink()
-    change_password(env)
+    if records_to_remove:
+        _logger.warning('Delete records from XML ID')
+        openupgrade.delete_records_safely_by_xml_id(env, records_to_remove)
+    if to_remove:
+        _logger.warning('Uninstalling not required modules')
+        modules_to_remove = env['ir.module.module'].search([
+            ('name', 'in', to_remove)])
+        modules_to_remove += modules_to_remove.downstream_dependencies()
+        modules_to_remove.module_uninstall()
+        modules_to_remove.unlink()
     env.cr.execute("""
         UPDATE ir_module_module
         SET
